move doGet and doPost to common
