+++ cmd/agent/app/status.go
@@ -51,7 +51,7 @@ func requestStatus() error {
 	c := GetClient()
 	urlstr := "http://" + sockname + "/agent/status"
 
-	r, e := doGet(c, urlstr)
+	r, e := common.DoGet(c, urlstr)
 	if e != nil {
 		var errMap = make(map[string]string)
 		json.Unmarshal(r, errMap)
